rename: 타입관련 파일 PascalCase 로 변경
