+++ packages/patchlogr-core/src/partition/types/HashNode.ts
