+++ packages/patchlogr-core/src/partition/types/HashObject.ts
